requirements.txt
agate==1.9.1
aiosqlite==0.22.1
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.7
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
babel==2.18.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
coolname==2.2.0
cryptography==46.0.5
daff==1.4.2
dateparser==1.2.2
dbt-adapters==1.17.3
dbt-common==1.34.1
dbt-core==1.10.19
dbt-duckdb==1.10.0
dbt-extractor==0.6.0
dbt-protos==1.0.431
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
dlt==1.21.0
docker==7.1.0
duckdb==1.4.4
exceptiongroup==1.3.1
fastapi==0.128.8
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
giturlparse==0.14.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
isodate==0.6.1
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.2.1
oauthlib==3.3.1
opentelemetry-api==1.39.1
orderly-set==5.5.0
orjson==3.11.5
packaging==25.0
parsedatetime==2.6
pathspec==0.12.1
pathvalidate==3.3.1
pendulum==3.2.0
pluggy==1.6.0
ply==3.11
prefect==3.4.25
prometheus_client==0.24.1
protobuf==6.33.5
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requirements-parser==0.13.0
rfc3339-validator==0.1.4
rich==14.3.2
rich-argparse==1.7.2
rpds-py==0.27.1
ruamel.yaml==0.19.1
semver==3.0.4
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.46
sqlglot==28.10.1
sqlparse==0.5.4
starlette==0.49.3
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tomli==2.4.0
tomlkit==0.14.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uv==0.10.2
uvicorn==0.39.0
websockets==15.0.1
zipp==3.23.0
